cats crud

+++ spycats/cat/models.py
@@ -4,9 +4,9 @@ from mission.models import Mission
 
 class Cat(models.Model):
     name = models.CharField(max_length=100)
-    years_of_experience = models.PositiveSmallIntegerField()
+    years_of_experience = models.SmallIntegerField()
     breed = models.CharField(max_length=100)
-    salary = models.PositiveIntegerField()
+    salary = models.SmallIntegerField()
     active_mission = models.OneToOneField(
         Mission,
         on_delete=models.SET_NULL,
